requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
# Pillow==8.3.1
email-validator==1.1.3
pyotp==2.6.0
python-dotenv==0.19.0
requests==2.26.0